Add 'exclude_pix' option to sqw.copy
This option allows one to make a copy of an SQW object without copying
the pixels. This may be beneficial as copying of pixels can be
expensive. For example the mask routine (yet to be implemented) will
build a new PixelData object from a subset of the original object's
pixels. In this case it is more efficient to not copy the original sqw
object's pixels.